requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
asyncpg==0.29.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
jsonschema==4.23.0
pgvector==0.3.2
numpy==1.26.4
orjson==3.10.7

jinja2==3.1.4
